// src/app/Models/employee.ts
export interface Employee {
  name: string;
  email: string;
  phone: string;
  address: string;
  gender: string;
  dateOfBirth: string;
  nationality: string;
  nationalId: string;
  hireDate: string;
  salary: number;
  attendanceTime: string;
  departureTime: string;
  isDeleted: boolean;
  departmentId: number;
  departmentName: string;
}